requirements.txt
anyio==3.6.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
fastapi==0.79.0
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
h11==0.13.0
httplib2==0.20.4
httptools==0.4.0
idna==3.3
oauthlib==3.2.0
packaging==21.3
protobuf==4.21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
uvicorn==0.18.2
watchfiles==0.16.0
websockets==10.3
